correct class,model, and migration names
