+++ app/views/texts/new.html.erb
@@ -1,4 +1,4 @@
-<%= form_for @phone, url: 'send_text' do |f| %>
+<%= form_for @text, url: 'send_text' do |f| %>
   <%= f.label :number %>
   <%= f.telephone_field :number %>
   <%= f.submit 'Send Text' %>
