+++ db/migrate/20170701160215_create_messages.rb
@@ -3,6 +3,7 @@ class CreateMessages < ActiveRecord::Migration[5.1]
     create_table :messages do |t|
       t.string :text
 
+
       t.timestamps
     end
   end
