--- db/migrate/20170701184542_create_phones.rb
@@ -1,9 +1,0 @@
-class CreatePhones < ActiveRecord::Migration[5.1]
-  def change
-    create_table :phones do |t|
-      t.string :number
-
-      t.timestamps
-    end
-  end
-end
