+++ db/migrate/20170701184542_create_texts.rb
@@ -1,0 +1,12 @@
+class CreateTexts < ActiveRecord::Migration[5.1]
+  def change
+    create_table :texts do |t|
+      t.string :number
+      t.string :body
+      t.string :description
+      t.datetime :send_at
+
+      t.timestamps
+    end
+  end
+end
